fix: position adjuster
