+++ .vscode/numbered-bookmarks.json
@@ -46,7 +46,7 @@
 			]
 		},
 		{
-			"path": "src/qubx/core/helpers.py",
+			"path": "src/qubx/core/strategy.py",
 			"bookmarks": [
 				{
 					"line": -1,
@@ -57,12 +57,12 @@
 					"column": 0
 				},
 				{
-					"line": 270,
-					"column": 49
-				},
-				{
 					"line": -1,
 					"column": 0
+				},
+				{
+					"line": 179,
+					"column": 11
 				},
 				{
 					"line": -1,
@@ -91,19 +91,19 @@
 			]
 		},
 		{
-			"path": "src/qubx/core/strategy.py",
+			"path": "src/qubx/utils/runner.py",
 			"bookmarks": [
+				{
+					"line": 144,
+					"column": 16
+				},
 				{
 					"line": -1,
 					"column": 0
 				},
 				{
-					"line": 322,
-					"column": 40
-				},
-				{
-					"line": 494,
-					"column": 31
+					"line": -1,
+					"column": 0
 				},
 				{
 					"line": -1,
