+++ src/qubx/core/basics.py
@@ -17,6 +17,28 @@ td_64 = np.timedelta64
 
 
 @dataclass
+class Signal:
+    """
+    Class for presenting signals generated by strategy
+    """
+
+    instrument: "Instrument"
+    signal: float
+    price: float | None = None
+    stop: float | None = None
+    take: float | None = None
+    group: str = ""
+    comment: str = ""
+
+    def __str__(self) -> str:
+        _p = f" @ { self.price }" if self.price is not None else ""
+        _s = f" stop: { self.stop }" if self.stop is not None else ""
+        _t = f" take: { self.take }" if self.take is not None else ""
+        _c = f" [{self.comment}]" if self.take is not None else ""
+        return f"{self.group} {self.signal:+f} {self.instrument.symbol}{_p}{_s}{_t} on {self.instrument.exchange}{_c}"
+
+
+@dataclass
 class FuturesInfo:
     contract_type: Optional[str] = None  # contract type
     delivery_date: Optional[datetime] = None  # delivery date
@@ -67,23 +89,19 @@ class Instrument:
             self._size_precision = int(abs(np.log10(self.min_size_step)))
         return self._size_precision
 
+    def signal(
+        self,
+        signal: float,
+        price: float | None = None,
+        stop: float | None = None,
+        take: float | None = None,
+        group: str = "",
+        comment: str = "",
+    ) -> Signal:
+        return Signal(self, signal, price, stop, take, group, comment)
+
     def __str__(self) -> str:
         return f"{self.exchange}:{self.symbol} [{self.market_type} {str(self.futures_info) if self.futures_info else 'SPOT ' + self.base + '/' + self.quote }]"
-
-
-@dataclass
-class Signal:
-    """
-    Class for presenting signals generated by strategy
-    """
-
-    instrument: Instrument
-    signal: float
-    price: Optional[float] = None
-    stop: Optional[float] = None
-    take: Optional[float] = None
-    group: Optional[str] = None
-    comment: Optional[str] = None
 
 
 class TransactionCostsCalculator:
@@ -128,6 +146,18 @@ class TransactionCostsCalculator:
 
 
 ZERO_COSTS = TransactionCostsCalculator("Zero", 0.0, 0.0)
+
+
+@dataclass
+class TriggerEvent:
+    """
+    Event data for strategy trigger
+    """
+
+    time: dt_64
+    type: str
+    instrument: Optional[Instrument]
+    data: Optional[Any]
 
 
 @dataclass
@@ -405,6 +435,16 @@ class SimulatedCtrlChannel(CtrlChannel):
 
     def receive(self) -> Any:
         raise ValueError("This method should not be called in a simulated environment.")
+
+
+class IComminucationManager:
+    databus: CtrlChannel
+
+    def get_communication_channel(self) -> CtrlChannel:
+        return self.databus
+
+    def set_communication_channel(self, channel: CtrlChannel):
+        self.databus = channel
 
 
 class ITimeProvider:
